[kahi_scienti_works] working in no doi support
